backend/requirements.txt
Django>=4.2.11
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
channels[daphne]>=4.1.0
channels-redis>=4.2.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
redis>=5.0.3
celery>=5.3.6
Pillow>=10.3.0
python-decouple>=3.8
opencv-python==4.9.0.80
ultralytics==8.2.18
mediapipe>=0.10.30,<0.11
roboflow>=1.1.9
